pipeline test helper class
